Bugfix: Finalization of TRTTIContext caused errors on process termination
Since the reference count of the global TRTTIContext instance was only
increasing the instance was never correctly 'freed'. This caused
exceptions if a process using DelphiJSON would end.
Since keeping at least one reference to the global TRTTIContext was
necessary to circumvent a multithreading bug/issue, we now keep one
instance at unit level and 'free' that instance in the 'finalization'
part of the unit.