Improved DJFromJSONFunctionAttribute by adding doNotInherit
If [doNotInherit] is set to true (the default value), class instances
that inherit from the class that defines the function, will not cause
the deserializer to call this function.

This allows patterns like Deserializing an abstract class into different
implementing classes using case distinction in the FromJSON method of
the abstract class, while being able to use a regular DelphiJSON<T> call
to deserialize the specific implementations.